+++ CHANGELOG.md
@@ -60,6 +60,7 @@ Bug fixes:
 * `IO#read(n)` no longer buffers more than needed, which could cause hanging if detecting readability via a native call such as `select(2)` (#1951).
 * Fixed `Random::DEFAULT.seed` to be different on boot (#1965, @kipply)
 * `rb_encoding->name` can now be read even if the `rb_encoding` is stored in native memory.
+* Detect and cut off recursion when inspecting a foreign object, substituting an ellipsis instead.
 
 Compatibility: